refactor: split account and invitation administration
