+++ crates/connect-hosted-provider/src/http.rs
@@ -37,13 +37,15 @@ use crate::{
     error::{ApiError, ApiResult},
     provider::{
         validate_limit, AuthorityRequestProof, HostedProvider, PrepareAuthorityImport,
-        PrepareAuthorityTransfer, ProviderAccountLimits, RegisterReplica, UpdateApplicationReplica,
+        PrepareAuthorityTransfer, RegisterReplica, UpdateApplicationReplica,
     },
 };
 
+mod accounts;
 mod authority_import_files;
 mod files;
 
+use accounts::account_routes;
 use authority_import_files::{
     commit_authority_import_file_upload, open_authority_import_file_upload,
     prepare_authority_import_file_part,
@@ -88,21 +90,6 @@ impl AppState {
             ))
         }
     }
-}
-
-#[derive(Debug, Deserialize)]
-struct CreateCollectionRequest {
-    account_id: Uuid,
-    collection_id: Uuid,
-    template: String,
-    display_name: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct UpsertAccountRequest {
-    entitlement_revision: u64,
-    #[serde(flatten)]
-    limits: ProviderAccountLimits,
 }
 
 #[derive(Debug, Deserialize)]
@@ -188,15 +175,7 @@ pub fn app(state: AppState) -> Router {
         .allow_methods([Method::GET, Method::POST, Method::DELETE])
         .max_age(std::time::Duration::from_secs(600));
     let internal = Router::new()
-        .route(
-            "/internal/v1/accounts/{account_id}",
-            put(upsert_account).get(account_usage),
-        )
-        .route(
-            "/internal/v1/accounts/{account_id}/collections/{collection_id}",
-            put(reconcile_collection_account),
-        )
-        .route("/internal/v1/collections", post(create_collection))
+        .merge(account_routes())
         .route(
             "/internal/v1/collections/{collection_id}",
             patch(rename_collection).delete(delete_collection),
@@ -374,58 +353,6 @@ async fn ready(State(state): State<AppState>) -> ApiResult<Json<Value>> {
         "status": "ready",
         "notifications": notifications
     })))
-}
-
-async fn create_collection(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(input): Json<CreateCollectionRequest>,
-) -> ApiResult<(StatusCode, Json<Value>)> {
-    state.authorize_internal(&headers)?;
-    let collection = state
-        .provider
-        .create_collection(
-            input.account_id,
-            input.collection_id,
-            &input.template,
-            &input.display_name,
-        )
-        .await?;
-    Ok((
-        StatusCode::CREATED,
-        Json(json!({ "collection": collection })),
-    ))
-}
-
-async fn upsert_account(
-    State(state): State<AppState>,
-    Path(account_id): Path<Uuid>,
-    Json(input): Json<UpsertAccountRequest>,
-) -> ApiResult<Json<Value>> {
-    let usage = state
-        .provider
-        .upsert_account(account_id, input.entitlement_revision, input.limits)
-        .await?;
-    Ok(Json(json!({ "account": usage })))
-}
-
-async fn account_usage(
-    State(state): State<AppState>,
-    Path(account_id): Path<Uuid>,
-) -> ApiResult<Json<Value>> {
-    let usage = state.provider.account_usage(account_id).await?;
-    Ok(Json(json!({ "account": usage })))
-}
-
-async fn reconcile_collection_account(
-    State(state): State<AppState>,
-    Path((account_id, collection_id)): Path<(Uuid, Uuid)>,
-) -> ApiResult<StatusCode> {
-    state
-        .provider
-        .reconcile_collection_account(account_id, collection_id)
-        .await?;
-    Ok(StatusCode::NO_CONTENT)
 }
 
 async fn rename_collection(
@@ -1021,25 +948,5 @@ mod tests {
         assert!(bool::from(hash.ct_eq(&hash)));
         let other: [u8; 32] = Sha256::digest(b"another-long-test-token-that-is-different").into();
         assert!(!bool::from(hash.ct_eq(&other)));
-    }
-
-    #[test]
-    fn collection_creation_requires_a_display_name() {
-        assert!(
-            serde_json::from_value::<CreateCollectionRequest>(serde_json::json!({
-                "account_id": Uuid::new_v4(),
-                "collection_id": Uuid::new_v4(),
-                "template": "tasknotes"
-            }))
-            .is_err()
-        );
-        let input: CreateCollectionRequest = serde_json::from_value(serde_json::json!({
-            "account_id": Uuid::new_v4(),
-            "collection_id": Uuid::new_v4(),
-            "template": "mdbase",
-            "display_name": "Worklog"
-        }))
-        .unwrap();
-        assert_eq!(input.display_name, "Worklog");
     }
 }
