+++ services/server/src/instance-admin.ts
@@ -5,6 +5,14 @@ import type {
   DatabaseQueryable
 } from "./db.js";
 import { normalizeEmailAddress } from "./email-identity.js";
+import {
+  invitationStatusCondition,
+  invitationSummary,
+  type InvitationPageInput,
+  type InvitationRow
+} from "./instance-admin-invitations.js";
+
+export type { InvitationPageInput } from "./instance-admin-invitations.js";
 
 const DEFAULT_PAGE_SIZE = 25;
 const MAX_PAGE_SIZE = 100;
@@ -28,12 +36,6 @@ export interface UserPageInput {
   status?: "active" | "suspended";
 }
 
-export interface InvitationPageInput {
-  limit?: number;
-  cursor?: string;
-  status?: "active" | "accepted" | "revoked" | "expired";
-}
-
 export interface BetaAccessRequestPageInput {
   limit?: number;
   cursor?: string;
@@ -52,21 +54,6 @@ interface UserSummaryRow {
   email: string | null;
   name: string;
   suspended_at: Date | string | null;
-  created_at: Date | string;
-}
-
-interface InvitationRow {
-  id: string;
-  email: string;
-  created_by: string;
-  expires_at: Date | string;
-  accepted_at: Date | string | null;
-  revoked_at: Date | string | null;
-  revoked_by: string | null;
-  revocation_reason: string | null;
-  send_count: number | string;
-  last_sent_at: Date | string | null;
-  entitlement_profile: string | null;
   created_at: Date | string;
 }
 
@@ -809,46 +796,6 @@ async function countByUser(
   return new Map(
     result.rows.map((row) => [row.user_id, Number(row.count)])
   );
-}
-
-function invitationSummary(row: InvitationRow) {
-  return {
-    id: row.id,
-    email: row.email,
-    status: invitationStatus(row),
-    created_by: row.created_by,
-    created_at: iso(row.created_at),
-    expires_at: iso(row.expires_at),
-    accepted_at: nullableIso(row.accepted_at),
-    revoked_at: nullableIso(row.revoked_at),
-    revoked_by: row.revoked_by,
-    revocation_reason: row.revocation_reason,
-    send_count: Number(row.send_count),
-    last_sent_at: nullableIso(row.last_sent_at),
-    entitlement_profile: row.entitlement_profile
-  };
-}
-
-function invitationStatus(
-  row: Pick<InvitationRow, "accepted_at" | "revoked_at" | "expires_at">
-): "active" | "accepted" | "revoked" | "expired" {
-  if (row.accepted_at) return "accepted";
-  if (row.revoked_at) return "revoked";
-  if (new Date(row.expires_at).getTime() <= Date.now()) return "expired";
-  return "active";
-}
-
-function invitationStatusCondition(
-  status: NonNullable<InvitationPageInput["status"]>
-): string {
-  if (status === "accepted") return "accepted_at IS NOT NULL";
-  if (status === "revoked") {
-    return "accepted_at IS NULL AND revoked_at IS NOT NULL";
-  }
-  if (status === "expired") {
-    return "accepted_at IS NULL AND revoked_at IS NULL AND expires_at <= now()";
-  }
-  return "accepted_at IS NULL AND revoked_at IS NULL AND expires_at > now()";
 }
 
 function pageSize(value: number | undefined): number {
